refactor: extract benches to benches/ directory
