+++ Cargo.toml
@@ -20,10 +20,9 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs", "--html-in-header", ".cargo/katex-header.html"]
 
 [[bench]]
-name = "criterion"
+name = "bench_uint"
 harness = false
-path = "criterion.rs"
-required-features = ["bench"]
+path = "benches/bench.rs"
 
 [dependencies]
 ruint2-macro = { version = "1.0.3", path = "./ruint-macro" }
@@ -57,12 +56,11 @@ postgres-types = { version = "0.2.3", optional = true }
 # TODO: remove the feature once 0.7 is released: https://github.com/launchbadge/sqlx/issues/1627
 sqlx-core = { version = "0.6", optional = true, features = ["runtime-tokio-native-tls"] }
 
-# Secretly dev dependencies
-# TODO: Move benchmarks to benches/
-criterion = { version = "0.5", optional = true }
-
 [dev-dependencies]
 ruint2 = { path = ".", features = ["arbitrary", "proptest"] }
+
+criterion = "0.5"
+rand = "0.8"
 
 approx = "0.5.1"
 ark-bn254 = "0.4.0"
@@ -99,6 +97,3 @@ zeroize = ["dep:zeroize"]
 
 postgres = ["dep:postgres-types", "dep:bytes"]
 sqlx = ["dep:sqlx-core"]
-
-# TODO: Move benchmarks to benches/
-bench = ["dep:criterion", "proptest", "rand"]
