+++ benches/benches/algorithms/div/mod.rs
@@ -1,0 +1,7 @@
+mod reciprocal;
+mod small;
+
+pub fn group(criterion: &mut criterion::Criterion) {
+    reciprocal::group(criterion);
+    small::group(criterion);
+}
