+++ src/algorithms/gcd/mod.rs
@@ -250,14 +250,3 @@ mod tests {
         });
     }
 }
-
-#[cfg(feature = "bench")]
-#[doc(hidden)]
-pub mod bench {
-    use super::*;
-    use criterion::Criterion;
-
-    pub fn group(criterion: &mut Criterion) {
-        matrix::bench::group(criterion);
-    }
-}
