+++ src/utils.rs
@@ -1,6 +1,6 @@
 /// Like `a % b` but returns `b` instead of `0`.
 #[must_use]
-pub const fn rem_up(a: usize, b: usize) -> usize {
+pub(crate) const fn rem_up(a: usize, b: usize) -> usize {
     let rem = a % b;
     if rem > 0 {
         rem
@@ -10,14 +10,14 @@ pub const fn rem_up(a: usize, b: usize) -> usize {
 }
 
 #[must_use]
-pub fn trim_end_slice<'a, T: PartialEq>(slice: &'a [T], value: &T) -> &'a [T] {
+pub(crate) fn trim_end_slice<'a, T: PartialEq>(slice: &'a [T], value: &T) -> &'a [T] {
     slice
         .iter()
         .rposition(|b| b != value)
         .map_or_else(|| &slice[..0], |len| &slice[..=len])
 }
 
-pub fn trim_end_vec<T: PartialEq>(vec: &mut Vec<T>, value: &T) {
+pub(crate) fn trim_end_vec<T: PartialEq>(vec: &mut Vec<T>, value: &T) {
     if let Some(last) = vec.iter().rposition(|b| b != value) {
         vec.truncate(last + 1);
     } else {
@@ -27,7 +27,7 @@ pub fn trim_end_vec<T: PartialEq>(vec: &mut Vec<T>, value: &T) {
 
 // Branch prediction hints.
 #[cfg(feature = "nightly")]
-pub use core::intrinsics::{likely, unlikely};
+pub(crate) use core::intrinsics::{likely, unlikely};
 
 // On stable we can use #[cold] to get a equivalent effect: this attribute
 // suggests that the function is unlikely to be called
@@ -38,7 +38,7 @@ fn cold() {}
 
 #[cfg(not(feature = "nightly"))]
 #[inline(always)]
-pub fn likely(b: bool) -> bool {
+pub(crate) fn likely(b: bool) -> bool {
     if !b {
         cold();
     }
@@ -47,7 +47,7 @@ pub fn likely(b: bool) -> bool {
 
 #[cfg(not(feature = "nightly"))]
 #[inline(always)]
-pub fn unlikely(b: bool) -> bool {
+pub(crate) fn unlikely(b: bool) -> bool {
     if b {
         cold();
     }
